feat: MVVM+SwiftUIバージョンでの単体テストの実装
#1
